Fix integer overflow in 16 bit compile
Only dosvga can compile as 16 bit and also show full color, and so this
bug has gone unnoticed.

+++ demos/testcurs.c
@@ -1209,7 +1209,7 @@ void gradient(int tmarg)
         move(tmarg + 3 + i, (COLS - 69) / 2);
         for (j = 0; j < len; j++)
         {
-            const int oval = j * 1000 / len;
+            const int oval = (int)(j * 1000L / len);
             const int reverse = 1000 - oval;
 
             if (!i)
